Revert "GF Signup:  Refactor models for better separation of concerns (#81644)"
This reverts commit d16e13eaa3a752624575e9976ca43887d33ff70e.
