--- client/my-sites/plans-features-main/hooks/use-is-free-domain-free-plan-upsell-enabled.ts
@@ -1,19 +1,0 @@
-import type { DataResponse } from 'calypso/my-sites/plans-grid/types';
-
-const useIsFreeDomainFreePlanUpsellEnabled = (
-	flowName?: string | null,
-	paidDomainName?: string | null
-): DataResponse< boolean > => {
-	if ( ! paidDomainName && ( flowName === 'onboarding' || flowName === 'onboarding-pm' ) ) {
-		return {
-			isLoading: false,
-			result: true,
-		};
-	}
-	return {
-		isLoading: false,
-		result: false,
-	};
-};
-
-export default useIsFreeDomainFreePlanUpsellEnabled;
