+++ client/my-sites/plans-features-main/hooks/use-is-plan-upsell-enabled-on-free-domain.ts
@@ -1,0 +1,18 @@
+import type { DataResponse } from 'calypso/my-sites/plans-grid/types';
+
+const useIsPlanUpsellEnabledOnFreeDomain = (
+	flowName?: string | null
+): DataResponse< boolean > => {
+	if ( flowName === 'onboarding' || flowName === 'onboarding-pm' ) {
+		return {
+			isLoading: false,
+			result: true,
+		};
+	}
+	return {
+		isLoading: false,
+		result: false,
+	};
+};
+
+export default useIsPlanUpsellEnabledOnFreeDomain;
