+++ client/my-sites/plans-features-main/index.tsx
@@ -13,7 +13,6 @@ import {
 	PLAN_HOSTING_TRIAL_MONTHLY,
 	PLAN_WOOEXPRESS_PLUS,
 	PLAN_ENTERPRISE_GRID_WPCOM,
-	PLAN_FREE,
 } from '@automattic/calypso-products';
 import { Button, Spinner } from '@automattic/components';
 import { WpcomPlansUI } from '@automattic/data-stores';
@@ -47,7 +46,6 @@ import scrollIntoViewport from 'calypso/lib/scroll-into-viewport';
 import { addQueryArgs } from 'calypso/lib/url';
 import PlanNotice from 'calypso/my-sites/plans-features-main/components/plan-notice';
 import PlanTypeSelector from 'calypso/my-sites/plans-features-main/components/plan-type-selector';
-import useIsFreePlanCustomDomainUpsellEnabled from 'calypso/my-sites/plans-features-main/hooks/use-is-free-plan-custom-domain-upsell-enabled';
 import { FeaturesGrid, ComparisonGrid } from 'calypso/my-sites/plans-grid';
 import useGridPlans from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-grid-plans';
 import usePlanFeaturesForGridPlans from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-plan-features-for-grid-plans';
@@ -62,14 +60,15 @@ import { getCurrentPlan, isCurrentUserCurrentPlanOwner } from 'calypso/state/sit
 import { getSitePlanSlug, getSiteSlug, isCurrentPlanPaid } from 'calypso/state/sites/selectors';
 import useStorageAddOns from '../add-ons/hooks/use-storage-add-ons';
 import ComparisonGridToggle from './components/comparison-grid-toggle';
+import { FreePlanFreeDomainDialog } from './components/free-plan-free-domain-dialog';
+import { FreePlanPaidDomainDialog } from './components/free-plan-paid-domain-dialog';
 import { LoadingPlaceHolder } from './components/loading-placeholder';
-import PlanUpsellModal from './components/plan-upsell-modal';
-import { useModalResolutionCallback } from './components/plan-upsell-modal/hooks/use-modal-resolution-callback';
 import usePricedAPIPlans from './hooks/data-store/use-priced-api-plans';
 import usePricingMetaForGridPlans from './hooks/data-store/use-pricing-meta-for-grid-plans';
 import useCurrentPlanManageHref from './hooks/use-current-plan-manage-href';
 import useFilterPlansForPlanFeatures from './hooks/use-filter-plans-for-plan-features';
-import useIsFreeDomainFreePlanUpsellEnabled from './hooks/use-is-free-domain-free-plan-upsell-enabled';
+import useIsCustomDomainAllowedOnFreePlan from './hooks/use-is-custom-domain-allowed-on-free-plan';
+import useIsPlanUpsellEnabledOnFreeDomain from './hooks/use-is-plan-upsell-enabled-on-free-domain';
 import useObservableForOdie from './hooks/use-observable-for-odie';
 import usePlanBillingPeriod from './hooks/use-plan-billing-period';
 import usePlanFromUpsells from './hooks/use-plan-from-upsells';
@@ -237,8 +236,8 @@ const PlansFeaturesMain = ( {
 	showLegacyStorageFeature = false,
 	isSpotlightOnCurrentPlan,
 }: PlansFeaturesMainProps ) => {
-	const [ isModalOpen, setIsModalOpen ] = useState( false );
-	const [ lastClickedPlan, setLastClickedPlan ] = useState< string | null >( null );
+	const [ isFreePlanPaidDomainDialogOpen, setIsFreePlanPaidDomainDialogOpen ] = useState( false );
+	const [ isFreePlanFreeDomainDialogOpen, setIsFreePlanFreeDomainDialogOpen ] = useState( false );
 	const [ showPlansComparisonGrid, setShowPlansComparisonGrid ] = useState( false );
 	const translate = useTranslate();
 	const storageAddOns = useStorageAddOns( { siteId, isInSignup } );
@@ -255,6 +254,11 @@ const PlansFeaturesMain = ( {
 		( state: IAppState ) => siteId && canUpgradeToPlan( state, siteId, PLAN_PERSONAL )
 	);
 	const previousRoute = useSelector( ( state: IAppState ) => getPreviousRoute( state ) );
+	const isCustomDomainAllowedOnFreePlan = useIsCustomDomainAllowedOnFreePlan(
+		flowName,
+		!! paidDomainName
+	);
+	const isPlanUpsellEnabledOnFreeDomain = useIsPlanUpsellEnabledOnFreeDomain( flowName );
 	const { setShowDomainUpsellDialog } = useDispatch( WpcomPlansUI.store );
 	const domainFromHomeUpsellFlow = useSelector( getDomainFromHomeUpsellInQuery );
 	const showUpgradeableStorage = config.isEnabled( 'plans/upgradeable-storage' );
@@ -265,19 +269,6 @@ const PlansFeaturesMain = ( {
 			? ! isCurrentPlanPaid( state, siteId ) || isCurrentUserCurrentPlanOwner( state, siteId )
 			: null
 	);
-
-	const isCustomDomainAllowedOnFreePlan = useIsFreePlanCustomDomainUpsellEnabled(
-		flowName,
-		paidDomainName
-	);
-	const isPlanUpsellEnabledOnFreeDomain = useIsFreeDomainFreePlanUpsellEnabled(
-		flowName,
-		paidDomainName
-	);
-	const resolveModal = useModalResolutionCallback( {
-		isCustomDomainAllowedOnFreePlan,
-		isPlanUpsellEnabledOnFreeDomain,
-	} );
 
 	const toggleShowPlansComparisonGrid = () => {
 		setShowPlansComparisonGrid( ! showPlansComparisonGrid );
@@ -317,24 +308,40 @@ const PlansFeaturesMain = ( {
 		return false;
 	};
 
+	const toggleIsFreePlanPaidDomainDialogOpen = useCallback( () => {
+		setIsFreePlanPaidDomainDialogOpen( ! isFreePlanPaidDomainDialogOpen );
+	}, [ isFreePlanPaidDomainDialogOpen ] );
+
 	const handleUpgradeClick = useCallback(
 		( cartItems?: MinimalRequestCartProduct[] | null ) => {
 			const cartItemForPlan = getPlanCartItem( cartItems );
-			const { product_slug: planSlug = PLAN_FREE } = cartItemForPlan ?? {};
-			setLastClickedPlan( planSlug );
-			if ( isFreePlan( planSlug ) ) {
-				recordTracksEvent( 'calypso_signup_free_plan_click' );
-			}
-
-			const displayedModal = resolveModal( planSlug );
-			if ( displayedModal ) {
-				setIsModalOpen( true );
-				return;
-			}
-
 			const cartItemForStorageAddOn = cartItems?.find(
 				( items ) => items.product_slug === PRODUCT_1GB_SPACE
 			);
+			const planPath = cartItemForPlan?.product_slug
+				? getPlanPath( cartItemForPlan.product_slug )
+				: '';
+			const checkoutUrlWithArgs = addQueryArgs(
+				{ ...( withDiscount && { coupon: withDiscount } ) },
+				`/checkout/${ siteSlug }/${ planPath }`
+			);
+
+			// `cartItemForPlan` is empty if Free plan is selected. Show `FreePlanPaidDomainDialog`
+			// in that case and exit. `FreePlanPaidDomainDialog` takes over from there.
+			// It only applies to main onboarding flow and the paid media flow at the moment.
+			// Standardizing it or not is TBD; see Automattic/growth-foundations#63 and pdgrnI-2nV-p2#comment-4110 for relevant discussion.
+			if ( ! cartItemForPlan ) {
+				recordTracksEvent( 'calypso_signup_free_plan_click' );
+				if ( paidDomainName ) {
+					toggleIsFreePlanPaidDomainDialogOpen();
+					return;
+				}
+				if ( isPlanUpsellEnabledOnFreeDomain.result ) {
+					setIsFreePlanFreeDomainDialogOpen( true );
+					return;
+				}
+			}
+
 			if ( cartItemForStorageAddOn?.extra ) {
 				recordTracksEvent( 'calypso_signup_storage_add_on_upgrade_click', {
 					add_on_slug: cartItemForStorageAddOn.extra.feature_slug,
@@ -346,17 +353,16 @@ const PlansFeaturesMain = ( {
 				return;
 			}
 
-			const planPath = cartItemForPlan?.product_slug
-				? getPlanPath( cartItemForPlan.product_slug )
-				: '';
-			const checkoutUrlWithArgs = addQueryArgs(
-				{ ...( withDiscount && { coupon: withDiscount } ) },
-				`/checkout/${ siteSlug }/${ planPath }`
-			);
-
 			page( checkoutUrlWithArgs );
 		},
-		[ onUpgradeClick, resolveModal, siteSlug, withDiscount ]
+		[
+			isPlanUpsellEnabledOnFreeDomain.result,
+			onUpgradeClick,
+			paidDomainName,
+			siteSlug,
+			toggleIsFreePlanPaidDomainDialogOpen,
+			withDiscount,
+		]
 	);
 
 	const term = usePlanBillingPeriod( {
@@ -645,32 +651,57 @@ const PlansFeaturesMain = ( {
 			<QuerySites siteId={ siteId } />
 			<QuerySitePlans siteId={ siteId } />
 			<QueryActivePromotions />
-			<PlanUpsellModal
-				isModalOpen={ isModalOpen }
-				paidDomainName={ paidDomainName }
-				modalType={ resolveModal( lastClickedPlan ) }
-				wpcomFreeDomainSuggestion={ wpcomFreeDomainSuggestion }
-				onClose={ () => setIsModalOpen( false ) }
-				onFreePlanSelected={ ( isDomainRetained ) => {
-					if ( ! isDomainRetained ) {
-						removePaidDomain?.();
-					}
-					// Since this domain will not be available after it is selected, invalidate the cache.
-					invalidateDomainSuggestionCache();
-					wpcomFreeDomainSuggestion.result &&
-						setSiteUrlAsFreeDomainSuggestion?.( wpcomFreeDomainSuggestion.result );
-					onUpgradeClick?.( null );
-				} }
-				onPlanSelected={ ( planSlug ) => {
-					if ( ! signupFlowSubdomain && wpcomFreeDomainSuggestion.result ) {
-						setSiteUrlAsFreeDomainSuggestion?.( wpcomFreeDomainSuggestion.result );
-					}
-					invalidateDomainSuggestionCache();
-					const cartItemForPlan = getCartItemForPlan( planSlug );
-					const cartItems = cartItemForPlan ? [ cartItemForPlan ] : null;
-					onUpgradeClick?.( cartItems );
-				} }
-			/>
+			{ paidDomainName && isFreePlanPaidDomainDialogOpen && (
+				<FreePlanPaidDomainDialog
+					paidDomainName={ paidDomainName }
+					wpcomFreeDomainSuggestion={ wpcomFreeDomainSuggestion }
+					suggestedPlanSlug={ PLAN_PERSONAL }
+					isCustomDomainAllowedOnFreePlan={ isCustomDomainAllowedOnFreePlan }
+					onClose={ toggleIsFreePlanPaidDomainDialogOpen }
+					onFreePlanSelected={ () => {
+						if ( isCustomDomainAllowedOnFreePlan.isLoading ) {
+							return;
+						}
+
+						if ( ! isCustomDomainAllowedOnFreePlan.result ) {
+							removePaidDomain?.();
+						}
+						// Since this domain will not be available after it is selected, invalidate the cache.
+						invalidateDomainSuggestionCache();
+						wpcomFreeDomainSuggestion.result &&
+							setSiteUrlAsFreeDomainSuggestion?.( wpcomFreeDomainSuggestion.result );
+						onUpgradeClick?.( null );
+					} }
+					onPlanSelected={ () => {
+						const cartItemForPlan = getCartItemForPlan( PLAN_PERSONAL );
+						const cartItems = cartItemForPlan ? [ cartItemForPlan ] : null;
+						onUpgradeClick?.( cartItems );
+					} }
+				/>
+			) }
+			{ isFreePlanFreeDomainDialogOpen && (
+				<FreePlanFreeDomainDialog
+					suggestedPlanSlug={ PLAN_PERSONAL }
+					freeSubdomain={ resolvedSubdomainName }
+					onClose={ () => setIsFreePlanFreeDomainDialogOpen( false ) }
+					onFreePlanSelected={ () => {
+						if ( ! signupFlowSubdomain && wpcomFreeDomainSuggestion.result ) {
+							setSiteUrlAsFreeDomainSuggestion?.( wpcomFreeDomainSuggestion.result );
+						}
+						invalidateDomainSuggestionCache();
+						onUpgradeClick?.( null );
+					} }
+					onPlanSelected={ () => {
+						if ( ! signupFlowSubdomain && wpcomFreeDomainSuggestion.result ) {
+							setSiteUrlAsFreeDomainSuggestion?.( wpcomFreeDomainSuggestion.result );
+						}
+						invalidateDomainSuggestionCache();
+						const cartItemForPlan = getCartItemForPlan( PLAN_PERSONAL );
+						const cartItems = cartItemForPlan ? [ cartItemForPlan ] : null;
+						onUpgradeClick?.( cartItems );
+					} }
+				/>
+			) }
 			{ siteId && (
 				<PlanNotice
 					visiblePlans={ gridPlansForFeaturesGrid.map( ( gridPlan ) => gridPlan.planSlug ) }
